pyproject.toml
[project]
name = "granola-archiver"
version = "0.3.0"
description = "Automated system to archive Granola meeting transcripts to GitHub"
authors = [
    { name = "Anjor Kanekar" },
]
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "granola-client>=0.3.0",
    "python-dateutil>=2.9.0.post0",
]

[project.scripts]
archiver = "archiver.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.20",
    "black>=23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["archiver"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.20",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/anjor/granola-archiver"
Repository = "https://github.com/anjor/granola-archiver"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
